requirements.txt
django-admin
django
celery
dotenv
djangorestframework
django-cors-headers
pymysql
openai
PyPDF2
python-docx
google-generativeai
instaloader